Show how compiler should handle inner functions.
If an Elm inner function doesn't close on any outer
variables, it should be expressed that way in JS.
Otherwise things like Html.Lazy are not gonna be
able to memoize the function.

It's reasonable in Elm to put helper functions inside
the only function that uses them, just to declutter
the outer namespace a bit.

The workaround to the Html.Lazy bug is to put the
function at module level.